\name{gwebvis}
\alias{gwebvis}
\title{Interface to webvis device}
\description{
	This function uses the \pkg{webvis} package to create a device
}
\usage{
gwebvis(wv, handler=NULL, action=NULL, container = NULL, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wv}{webvis argument}
  \item{handler}{handler}
  \item{action}{action argument}
  \item{container}{parent container for processing object}
  \item{\dots}{not used}
}
\details{
  ???
}
\value{
  Returns a \pkg{proto} object with class gWebvis.
}
\references{The webvis device package}
\author{John Verzani }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\examples{
%\dontrun{}
%}
\keyword{interface}

